css/sidebar -move sidebar to component
