+++ client/src/components/Sidebar.css
@@ -1,0 +1,92 @@
+.header {
+  float: left;
+  height: 100%;
+  width: 120px;
+  position: fixed;
+  z-index: 1;
+  display: flex;
+  flex-direction: column;
+  padding: 10px 0;
+  background-color: #000000;
+}
+.header-logo {
+  font-size: 30px;
+  color: white;
+  margin: 0 auto;
+  padding: 20px 5px;
+  text-align: center;
+  border-bottom: 1px solid grey;
+}
+
+.header-element {
+  color: #ffffff;
+  text-transform: uppercase;
+  font-size: 12px;
+  font-weight: 700;
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 5px;
+  padding: 20px 5px;
+  margin: 0 auto;
+  border-bottom: 1px solid grey;
+  width: fit-content;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+.header-element:hover {
+  color: #ffffff;
+}
+
+.header-element--home::before {
+  content: url('../assets/img/home-icon.svg');
+  cursor: pointer;
+}
+
+.header-element--add::before {
+  content: url('../assets/img/add-icon.svg');
+  cursor: pointer;
+}
+
+.header-element--login::before {
+  content: url('../assets/img/user-icon.svg');
+  cursor: pointer;
+}
+
+.header-element--logout::before {
+  content: url('../assets/img/logout-icon.svg');
+  cursor: pointer;
+}
+
+.header .user {
+  /* position: absolute;
+    right: 0; */
+  font-weight: bold;
+  font-size: 18px;
+}
+
+.user_is-activated {
+  position: relative;
+  display: inline-block;
+}
+
+.user_info {
+  cursor: pointer;
+}
+
+.user_ava {
+  display: inline-block;
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+}
+
+.user_ava img {
+  vertical-align: middle;
+  width: 100%;
+  height: 100%;
+  border-radius: 50%;
+}
