+++ client/src/pages/Home/Home.css
@@ -30,153 +30,7 @@ body {
   height: 100%;
   z-index: -1;
 }
-.home .header {
-  float: left;
-  height: 100%;
-  width: 120px;
-  position: fixed;
-  z-index: 1;
-  display: flex;
-  flex-direction: column;
-  padding: 10px 0;
-  background-color: #000000;
-}
-.home .header-logo {
-  font-size: 30px;
-  color: white;
-  margin: 0 auto;
-  padding: 20px 5px;
-  text-align: center;
-  border-bottom: 1px solid grey;
-}
 
-.home .header-element {
-  color: #fff;
-  text-transform: uppercase;
-  font-size: 12px;
-  font-weight: 700;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 5px;
-  padding: 20px 5px;
-  margin: 0 auto;
-  border-bottom: 1px solid grey;
-  width: fit-content;
-}
-
-.home .header-element a {
-  color: #fff;
-  text-transform: uppercase;
-  font-size: 12px;
-  font-weight: 700;
-  text-align: center;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.home .header-element--home::before {
-  content: url('../../assets/img/home-icon.svg');
-  cursor: pointer;
-}
-
-.home .header-element--add::before {
-  content: url('../../assets/img/add-icon.svg');
-  cursor: pointer;
-}
-
-.home .header-element--login::before {
-  content: url('../../assets/img/user-icon.svg');
-  cursor: pointer;
-}
-
-.home .header-element--logout::before {
-  content: url('../../assets/img/logout-icon.svg');
-  cursor: pointer;
-}
-
-.home .header .user {
-  /* position: absolute;
-    right: 0; */
-  font-weight: bold;
-  font-size: 18px;
-}
-
-.home .user_is-activated {
-  position: relative;
-  display: inline-block;
-}
-
-.home .user_info {
-  cursor: pointer;
-}
-
-.home .user_ava {
-  display: inline-block;
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-}
-
-.home .user_ava img {
-  vertical-align: middle;
-  width: 100%;
-  height: 100%;
-  border-radius: 50%;
-}
-
-.home .user_name {
-  display: inline-block;
-  color: #000;
-  margin-left: 5px;
-}
-
-.home .user_logout {
-  display: none;
-  position: absolute;
-  background-color: #f9f9f9;
-  min-width: 100%;
-  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
-  padding: 12px 16px;
-  z-index: 1;
-}
-
-.home .user_is-activated:hover .user_logout {
-  display: block;
-}
-
-.home .user_logout button {
-  background: #fff;
-  width: 100%;
-  font-size: 12px;
-  line-height: 16px;
-  font-weight: 600;
-  text-transform: uppercase;
-  padding: 3px 0;
-}
-
-.home .user_login-link a {
-  width: auto;
-  color: rgba(19, 19, 233, 0.896);
-  margin: 0;
-  background-color: transparent;
-  border: none;
-  font-weight: 600;
-  text-decoration: none;
-}
-.home .button-add {
-  height: 70px;
-  width: 70px;
-  border-radius: 50% !important;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-  margin-left: auto;
-  margin-right: 10%;
-}
 .main-content {
   margin-left: 120px;
   width: calc(100% - 120px);
